[NNO-8494] У InputText не проставляется min,max для типа NUMERIC
(cherry picked from commit 8db47efff4c5f89fbb4f28bfb043431242956acb)

# Conflicts:
#	backend/n2o/n2o-config/src/test/resources/net/n2oapp/framework/config/metadata/compile/control/testInputText.page.xml

+++ backend/n2o/n2o-config/src/main/java/net/n2oapp/framework/config/metadata/compile/control/InputTextCompiler.java
@@ -1,19 +1,15 @@
 package net.n2oapp.framework.config.metadata.compile.control;
 
-import net.n2oapp.framework.api.metadata.ReduxModel;
 import net.n2oapp.framework.api.metadata.Source;
 import net.n2oapp.framework.api.metadata.compile.CompileContext;
 import net.n2oapp.framework.api.metadata.compile.CompileProcessor;
 import net.n2oapp.framework.api.metadata.control.plain.N2oInputText;
 import net.n2oapp.framework.api.metadata.domain.Domain;
-import net.n2oapp.framework.api.metadata.meta.ModelLink;
 import net.n2oapp.framework.api.metadata.meta.control.InputText;
 import net.n2oapp.framework.api.metadata.meta.control.StandardField;
-import net.n2oapp.framework.config.metadata.compile.context.PageContext;
-import net.n2oapp.framework.config.metadata.compile.page.PageScope;
-import net.n2oapp.framework.config.metadata.compile.widget.WidgetScope;
-import net.n2oapp.framework.config.util.CompileUtil;
 import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
 
 import static net.n2oapp.framework.api.metadata.compile.building.Placeholders.property;
 import static net.n2oapp.framework.api.metadata.local.util.CompileUtil.castDefault;
@@ -73,8 +69,8 @@ public class InputTextCompiler extends StandardFieldCompiler<InputText, N2oInput
                 inputText.setStep(castDefault(source.getStep(), "1"));
                 return;
             case NUMERIC:
-                inputText.setMin(p.cast(p.resolveJS(source.getMin(), Long.class), "-999999999"));
-                inputText.setMax(p.cast(p.resolveJS(source.getMax(), Long.class), "999999999"));
+                inputText.setMin(p.cast(p.resolveJS(source.getMin(), BigDecimal.class), "-999999999"));
+                inputText.setMax(p.cast(p.resolveJS(source.getMax(), BigDecimal.class), "999999999"));
                 inputText.setStep(castDefault(source.getStep(), "0.01"));
                 inputText.setPrecision(p.cast(source.getPrecision(), p.resolve(property("n2o.api.control.input.text.precision"), Integer.class)));
                 return;
